# doc/api.yaml
openapi: 3.0.0
paths:
...
  /track:
    get:
      operationId: TrackController_findAll
      summary: Get tracks list
      description: Gets all library tracks list
      parameters: []
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TrackEntity'
      tags: &ref_2
        - Track
    post:
      operationId: TrackController_create
      summary: Add new track
      description: Add new track information
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTrackDto'
      responses:
        '201':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TrackEntity'
        '400':
          description: Bad request. body does not contain required fields
      tags: *ref_2
  /track/{id}:
    get:
      operationId: TrackController_findOne
      summary: Get single track by id
      description: Gets single track by id
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TrackEntity'
        '400':
          description: Bad request. trackId is invalid (not uuid)
        '404':
          description: Track not found
      tags: *ref_2
    put:
      operationId: TrackController_update
      summary: Update track information
      description: Update library track information by UUID
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTrackDto'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TrackEntity'
        '400':
          description: Bad request. trackId is invalid (not uuid)
        '404':
          description: Track not found
      tags: *ref_2
    delete:
      operationId: TrackController_remove
      summary: Delete track
      description: Delete track from library
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '204':
          description: The track has been deleted
        '400':
          description: Bad request. trackId is invalid (not uuid)
        '404':
          description: Track not found
      tags: *ref_2
...
  /favs/track/{id}:
    post:
      operationId: FavsController_createTrack
      summary: Add track to the favorites
      description: Add track to the favorites
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '201':
          description: Added successfully
        '400':
          description: Bad request. trackId is invalid (not uuid)
        '422':
          description: Track does not exist
      tags: *ref_4
    delete:
      operationId: FavsController_deleteTrack
      summary: Delete track from favorites
      description: Delete track from favorites
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '204':
          description: Deleted successfully
        '400':
          description: Bad request. trackId is invalid (not uuid)
        '404':
          description: Track was not found
      tags: *ref_4
info:
  title: Home Library service
  description: Home Library service REST API by Pas92
  version: '1.0'
  contact: {}
...
servers: []
components:
  schemas:
...
    TrackEntity:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        artistId:
          type: string
          nullable: true
        albumId:
          type: string
          nullable: true
        duration:
          type: number
      required:
        - id
        - name
        - artistId
        - albumId
        - duration
    CreateTrackDto:
      type: object
      properties:
        name:
          type: string
        artistId:
          type: string
          nullable: true
        albumId:
          type: string
          nullable: true
        duration:
          type: number
      required:
        - name
        - artistId
        - albumId
        - duration
    UpdateTrackDto:
      type: object
      properties:
        name:
          type: string
        artistId:
          type: string
          nullable: true
        albumId:
          type: string
          nullable: true
        duration:
          type: number
...